Implement UPS Service with rate estimation
This implements the UPS' own rate estimation endpoint. It uses the XML
API (rather than the JSON or SOAP API they also offer). All requests are
POST requests.

All shipment and package options should be given as `options` for the
FriendlyShipping::Shipment (if they relate to the entire shipment) or as
`properties` on individual packages (if they relate to only one package
within a shipment, such as a reference number).