Add USPS rate estimation endpoint
This adds the capability of estimating USPS rates directly from USPS.
There is pretty complicated code, because the USPS rating API has a lot
of functionality and an interesting data model.

When requesting rates for a shipment, USPS will return all possible
rates for each package. This means that we need to match the response's
rates to each of the packages, and then sum the rates for all the
packages to get rates for the entire shipment.

Compared to previous API implementations, we're using USPS' endpoint
with POST requests rather than GET requests because that makes VCR
cassette matching easier and probably also allows longer a longer
request payload.

Co-Authored-By: Matthew Bass <matt@candlescience.com>