Add API Result and Failure container objects
These objects are intended to wrap results from our APIs, similar to
what AddressValidationResult does, but for all cases. The naming is now
much more generic, so that we can now easily wrap all calls. I've
dropped the idea of storing the original arguments that the service have
been called with, as the calling code will very likely know what
arguments it uses, while request and response are new.