Create generic HTTP client
All of the HTTP clients we have wrap a success response from
`rest-client` in a `Success` Monad and any errors in a `Failure` Monad.
Only the ShipEngine client has some special behaviour. This commit
extracts that special behaviour to a BadRequestHandler specific to
ShipEngine, and otherwise uses a simple private method.