Do not mark commercial addresses as residential for UPS
Previously, the UPS API would return a commercial rate if the
destination address was commercial, even if the residential flag was
included in the API call. That appears to no longer be the case. Now,
UPS will sometimes return a commercial rate if it knows the address
is commercial, but other times it won't. Therefore, we need to
conditionally include the residential flag based on the address type.

Co-Authored-By: Matthew Bass <matt@candlescience.com>