Extract shipping method to separate argument in labels API
When generating a label, at the time we generate it the shipping method
service code is not a property of the shipment (yet). I believe this is
a clearer API that reads better: Having a shipment, and then specifying
explicitly which shipping method to use. Bonus: That shipping method has
a `carrier` property which we can use rather than adding yet another
optional string.