Add Shipment Options class
Whenever we need to pass options to some services' API, we often need to
include options that tell us HOW we want to estimate rates, create
labels, etc. These options should be isolated from the shipment (WHAT we
want to ship) itself. Now shipments have packages, and the intended
functionality of this class is to support selecting options from a
separate class for the shipment's packages. This class is ALSO intended
to be subclassed in individual options class. A typical shipment options
class would look like this:

```rb
class RateEstimatesOptions < ShipmentOptions
	attr_reader :carriers

	def initialize(carriers: [], **kwargs)
		@carriers = carriers
		super kwargs
	end
end
```