Add changed address type to UPS rate estimates
This modifies the rate estimates UPS service. The returned rates now
include a new address type if it was changed by UPS before quoting the
rate. The new address type is added to the data hash. If the address
type is unchanged by UPS, it won't appear in the data hash.