Fix timing estimation for users from Alaska and Hawaii
When estimating timing to non-contiguous US states, we get one estimate for
USPS First Class Package Service that does not, in fact, contain a timing
estimate, but rather a friendly message that it's difficult to ship things
there. This commit stops our parser from breaking with a NoMethodError in these cases.