Add rectangular boolean to USPS rate options
Since the presence of the Girth element controls the treatment of a
package as non-rectangular, we need a new package option that specifies
when a package is rectangular so the Girth element can be excluded from
the request. A rectangular package's dimensional weight is calculated
differently than a non-rectangular package.