UPS Freight: Use current endpoint
The UPS Freight API has a new, more restful endpoint with better error
handling. Not that much changed between the old one and this one, except
that both request generation and response parsing becomes a little bit
simpler.

Additionally, with this, we can share the error handler and HTTP client
between the rates API and the upcoming labels/shipment API.