Combine Freight and default API error handlers
This combines the Freight and default API error handlers into one
handler that can accept an option API error class. This class is used
to wrap the underlying API exception and, in the case of the Freight
error handler, parses the response body JSON into a message.

In the future, API error classes can be customized for the other
carrier services to similarly parse response messages.