Refactor R+L BOL creation to return new object
This refactors the R+L BOL creation API method to return the new
`ShipmentInformation` object which contains the PRO number, pickup
request number, documents, etc.

The BOL and label printing API methods take this object as an argument
and append the appropriate document to it on success.

This makes it easier to chain these 2 or 3 method calls together while
only dealing with a single object containing the results.