Return array from ShipEngine `#validate_address`
For consistency with other service classes, we should be returning an
array of locations from this method instead of a single location.

Eventually, we may allow an array of locations to be passed into this
method, in which case returning an array makes more sense.

For now, our other service classes only accept a single location for
validation while returning an array as the result. This doesn't make a
lot of sense, but let's match it for consistency.