Remove duplicate package deprecation warnings
This cleans up some deprecation warnings that were duplicating
warnings elsewhere in the code.

For example, we were checking for `options.package_options` in some
places and warning about deprecations if the options existed, but this
would also trigger the warning within the `#package_options` method
itself which was unnecessary.