Refactor USPS Ship rates for multiple API calls
Unlike the old API, the new USPS API only returns rates for one
package at a time. This means we have to make separate API calls, one
for each package, and sum the rates to arrive at a total.

We accomplish this using the `Dry::Monads::Do` mixin which lets us
`yield` to each API call. Any failed API call will short circuit the
process and immediately return the failure.

If all API calls succeed, we sum the rates and return a new `Rate`
wrapped in `ApiResult` and a `Success` monad.