Fix TForce shipping label codes
The old TForce (UPS) API used 2 codes to determine whether the
shipping label was thermal or not. The new TForce API uses a wide
variety of label type codes for different label sizes and formats.
This commit adds the new codes to the document options class and
deprecates the old `thermal` attribute.