Never send empty UPS `ShipmentRatingOptions`
This fixes a bug when the UPS rates option `negotiated_rates` flag is
set to false. The `ShipmentRatingOptions` value ends up being an empty
hash. The UPS API used to tolerate this just fine, but in a recent
update they began complaining with an error response like this:

"Atleast one shipment rating option either negotiated or FRS rates or
ratechart indicator is required."

This changes the rates payload generator to leave out the
`ShipmentRatingOptions` key when `negotiated_rates` is false.