Deprecate the `ApiFailure` class
It's easy to confuse `ApiFailure` with `ApiError` and `ApiResult`. We
really only need one class to represent an API result, especially
since we use the `Dry::Monads::Result` library to indicate if a result
is success/failure. This commit deprecates `ApiFailure` and moves the
`#failure` attribute alias into `ApiResult` along with the `#to_s`
behavior which is to just return `data` as a string.

In a future release we can remove the `ApiFailure` class entirely.