Use Zeitwerk to load FriendlyShipping
Currently, FriendlyShipping is a pretty slow gem to load, because all
services and their shipping methods are loaded on app startup,
regardless if they're being used or not. This sets up Zeitwerk to
autoload constants in the project, allowing us to drop all require
statements to our own project in the codebase, and speeding up
require'ing this gem.