Add density eligible boolean to TForce BOL
Updates the TForce BOL options with a "density eligible" boolean that
can optionally be included in the API request to get a BOL.

When this flag is enabled, the rate will be based on pallet density
instead of NMFC code and Freight class. Dimensions and weight for each
pallet must be available for this to work.